requirements.txt
autopep8==1.5.5
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Pillow==8.1.0
pycodestyle==2.6.0
toml==0.10.2
Werkzeug==1.0.1
